Updates to remove flashing/disappearing modal fixes issues 26
I tried to do this is in the least damaging way possible that also allows for classes to be easily changed to suit the needs of the project. By limiting the modal size to not over run the screen, the modal will always be able to render on the viewport.

This will break for people who have published their classes, so I've included a list of the changes I've made that they can adapt. You Can also just look at the changed filed in the PR.

***Only do these if you published or overrode the classes from laravel-ui/modal***
1. Add 'maxWidthClass' => $componentClass::modalMaxWidthClass(), to the 'modelAttributes' in Modal Class
2. Add the protected static array $maxWidths to the abstract ModelComponent
3. copy the static function modalMaxWidthClass to abstract ModelComponent
4. In resources/js/modal.js, set the default class to null
5. In the same file replace this line `this.modalWidth = 'sm:max-w-' + this.getActiveComponentModalAttribute('maxWidth');` with this one `this.modalWidth = this.getActiveComponentModalAttribute('maxWidthClass');`twice.
6. Finally in the same file initialize the the default setting of maxWidth to your configured default in this way...

```js
in init() {

// add this line
this.modalWidth = this.getActiveComponentModalAttribute('maxWidthClass');

// before
this.$watch...

7. run npm run prod and you should publish a new modal.js that will now reflect those changes.

*** Remember that only if you have published some or all of your classes and configs otherwise this should just work ***

PR'd with ❤️
-Roni